apps/agent/pyproject.toml
[project]
name = "dos-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.6",
  "uvicorn[standard]==0.34.0",
  "psycopg[binary]==3.2.4",
  "psycopg-pool==3.2.4",
  "openai>=2,<3",
  "httpx==0.28.1",
  "pymupdf==1.26.4",
  "pyyaml==6.0.2",
  "livekit==1.1.2",
  "livekit-agents==1.4.3",
  "livekit-plugins-deepgram==1.4.3",
  "livekit-plugins-openai==1.4.3",
  "livekit-plugins-silero==1.4.3",
  "python-dotenv==1.0.1",
  "stripe>=11,<12",
  "icalendar==6.1.3"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
test = [
  "pytest>=8,<9",
  "pytest-asyncio>=0.24,<1",
  "httpx>=0.28,<1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
